Fix exports and rename SchemaRelationshipDiagram
