+++ packages/ui/stories/viewer-table/DictionaryHeader.stories.tsx
@@ -3,7 +3,6 @@
 // import { DictionaryHeader } from '#/viewer-table/DictionaryHeader';
 import type { Meta, StoryObj } from '@storybook/react';
 import { pick } from 'lodash';
-import type { PartialTheme } from '../../src/theme';
 import { DictionaryHeader } from '../../src/viewer-table/DictionaryHeader';
 import biosampleDictionary from '../fixtures/minimalBiosampleModel';
 import themeDecorator from '../themeDecorator';
